SmartPicture-backend/requirements.txt
# === SmartPicture Backend Requirements ===
# 基础 Web 框架
Flask>=3.0.0

# === Google AI 核心依赖 ===
google-generativeai>=0.5.0
google-cloud-aiplatform>=1.49.0

# === Firebase / Firestore 支持 ===
google-cloud-firestore>=2.17.0
firebase-admin>=6.5.0

# === 生产部署 ===
gunicorn

# === 可选扩展功能（按需安装） ===
# google-cloud-vision         # 图像内容分析
# google-cloud-speech         # 语音转文本
# google-cloud-texttospeech   # 文本转语音
# google-cloud-storage        # 文件存储
# google-cloud-pubsub         # 异步任务队列
# google-cloud-bigquery       # 数据分析
# google-cloud-translate      # 多语言翻译
# Pillow                      # 图像处理（SmartPicture常用）
# requests                    # 外部 API 调用
